Copied functions to c file, filled the header file, added test txt file
