+++ telephoneDirectory/directory.txt
@@ -1,0 +1,4 @@
+Ivan Ivanov
+89100000000
+Nikolay Zaytsev
+82109467284
